requirements.txt
gunicorn

Flask
Jinja2
requests
SQLAlchemy
Werkzeug
Flask-Babel
psycopg2
Flask-Session
python-dotenv
base64
imagekitio